Projectile direction, character facing direction, gameobject hierarchy
The projectile now launches towards the direction the player is facing.

The character now faces the direction of its movement.

Changed the gameobject hierarchy in the scene to simplify it.

Changed the name of the scene to "Level 1" instead of "SampleScene" to avoid problems in the future.

Also updated the prefabs to match the ones used in the scene.