Quick base for health system with health indicators in the UI
Placed health indicators on the UI.

The system can find existing health indicators and then disable the correct amount of them if the player has already lost health in a previous level.

The current health is always stored in playerprefs and the playerpref and healthindicators are reset if the current level is the first one.